Implement end-game conditions
The win-condition check if the cells are the same as the recently marked
cell in the same row, column, or any of the two axial diagonals of
the board.

The implementation of the win condition required the use of a new method
(isHomogeneousCellGroup) of the Gameboard object that returns whether or
not all the cells in a group contains the same mark. A group contains
cells in a line from one boundary to its opposite of the board. The
orientation of the line is determined by passing callback functions
as the coordinates of the cell that may return a constant value or a
linear relationship.

The tie condition merely determines whether all cells have been marked.
However, it is not explicitly implemented by the code, as the code only
checks if the round number is the same as the size of the board. This
makes sense because after that number of rounds, the players will
have no more cells to mark. Hence, assuming that no one has won, the tie
condition evaluates to true.

The current limitations to the code include:
    - the win and tie conditions are decoupled and accessible by
      external code. This means that the responsibility to test them
      in the correct order lies on the user of their respective methods.

    - as a corollary to the last point, the parameters of the win method
      that must correspond to the cell coordinates of the recently-added
      mark necessitates the user to maintain consistency between the two
      in order for the end-game logic to function per requirements.

    - The tie condition is inefficient as it is not able to predict
      its outcome until board is completely filled.